pyproject.toml
[project]
name = "deploy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
